move deprecation decorator to new _internal.py module
